Fix the purple and green textures
I realized that, when changing the color of the progression item, I just selected everything that was that purple color and replaced it with green. This green not only didn't look great, it also could have just been color-picked from the green in the default AP logo texture. So I returned all of the pixels that were replaced with green to their original color and I replaced the green with the color picked version.