onLevelLoaded needed to be called earlier
specifically, before we notify that the settings are available
this is because of two reasons:
* we may want to change the gamestate as a result of the options being present
* onLevelLoaded sets a lot of variables that are undefined beforehand (like localCheckedLocations)

i discovered this bug as it was a combination of area-setting and new-game-ap side effects.
new-game-ap set the gamestate to running instead of title as a result of the options having loaded.
then, area-setting tried to send the area message out despite the fact that it should have been gated off by isTitle.
this would have probably caused problems with the newgame patch, so i'm glad i caught it when i did.